+++ backend/src/main/java/sg/backend/controller/FileController.java
@@ -1,10 +1,8 @@
 package sg.backend.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.io.Resource;
-import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,7 +12,7 @@ import sg.backend.service.FileService;
 @RestController
 @RequestMapping("/file")
 @RequiredArgsConstructor
-public class FileApiController {
+public class FileController {
 
     private final FileService fileService;
