+++ backend/src/main/java/sg/backend/service/FileService.java
@@ -1,7 +1,26 @@
 package sg.backend.service;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.Resource;
+import org.springframework.core.io.UrlResource;
+import org.springframework.stereotype.Service;
 
-public interface FileService {
-    Resource getImage(String fileName);
+@Service
+public class FileService {
+
+    @Value("${file.path}")
+    private String filePath;
+
+    public Resource getImage(String fileName) {
+        Resource resource = null;
+
+        try {
+            resource = new UrlResource("file:" + filePath + fileName);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
+        }
+
+        return resource;
+    }
 }
