--- backend/src/main/java/sg/backend/service/implement/FileServiceImpl.java
@@ -1,28 +1,0 @@
-package sg.backend.service.implement;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.io.Resource;
-import org.springframework.core.io.UrlResource;
-import org.springframework.stereotype.Service;
-import sg.backend.service.FileService;
-
-@Service
-public class FileServiceImpl implements FileService {
-
-    @Value("${file.path}")
-    private String filePath;
-
-    @Override
-    public Resource getImage(String fileName) {
-        Resource resource = null;
-
-        try {
-            resource = new UrlResource("file:" + filePath + fileName);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return null;
-        }
-
-        return resource;
-    }
-}
